requirements.txt
anyio==3.6.2
certifi==2023.5.7
click==8.1.3
dnspython==2.3.0
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
motor==3.1.2
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==0.21.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.0
uvicorn==0.22.0
zipp==3.15.0
